Выполнил задание: время разбивать камни
