+++ public/index.html
@@ -145,6 +145,6 @@
       </section>
     </main>
 
-    <script src="./js/main.js"></script>
+    <script src="./bundle.js"></script>
   </body>
 </html>
